Move file-tree-store.ts to file-explorer/controllers
